pyproject.toml
[tool.poetry]
name = "openfoodfacts"
version = "2.2.0"
description = "Official Python SDK of Open Food Facts"
authors = ["The Open Food Facts team"]
license = "Apache 2.0"
readme = "README.md"

[tool.mypy]
ignore_missing_imports = true

[tool.isort] # From https://black.readthedocs.io/en/stable/compatible_configs.html#isort
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
requests = ">=2.20.0"
pydantic = ">=2.0.0,<3.0.0"
tqdm = ">=4.0.0,<5.0.0"
redis = { version = "~5.1.0", optional = true, extras = ["hiredis"] }
Pillow = { version = ">=9.3,<10.4", optional = true }

[tool.poetry.group.dev.dependencies]
requests-mock = "1.11.0"
flake8 = "6.0.0"
black = "24.3.0"
mypy = "1.12.0"
isort = "5.12.0"
coverage = {version = "7.2.7", extras = ["toml"]}
pytest = "7.4.2"
types-requests = "2.31.0.2"
types-tqdm = "4.66.0.20240417"
types-redis = "^4.6.0.20240425"


[tool.poetry.group.ml.dependencies]
tritonclient = {extras = ["grpc"], version = "^2.52.0"}
opencv-python = "^4.10.0.84"

[tool.poetry.extras]
redis = ["redis"]
Pillow = ["Pillow"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
